nedd some fixes

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mbousbaa <mbousbaa@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 05:35:19 by mbousbaa          #+#    #+#             */
-/*   Updated: 2022/10/26 17:42:19 by mbousbaa         ###   ########.fr       */
+/*   Updated: 2022/10/27 00:52:13 by mbousbaa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ size_t	ft_strlen(char *s);
 void	ft_putchar(char c, size_t *count);
 void	ft_putstr(char *str, size_t *count);
 void	ft_putnbr(int nb, size_t *count);
-void	ft_putnbr_base(unsigned int nb, char *base, size_t *count);
+void	ft_putnbr_base(unsigned long nb, char *base, size_t *count);
 int		ft_printf(const char *str, ...);
 
 #endif
